feat(discriminators): for types, validation and logging
Could be considered a "chore" since in theory, there are no user
facing changes.  Rather, we now allow a "@discriminator quoteType"
JSDoc decorator, and use the non-standard OpenAPI json-schema
extension "discriminator" to support this with `oneOf`.

The main advantage here is that now for a mistmatch on e.g. Quote,
we can only check against the correct subschema (ETF, equity, etc)
rather than logging ALL the multitude of non-matches for all subs.

This makes it easier to understand where the actual problem is,
rather then needing to look through all the errors and figure out
the relevant part.