feat(npm/cli): "bin" field in package.json / converted npm package
This was a staple of v2 and unfortunately disappeared with the
initial npm conversion.  Now the functionality is restored.

Outside of npm, "deno task cli" uses the local package, but we
should still explore global installation options.

Also very easy with Deno now to offer precompiled binaries for the
popular OSes.