fix(bundling): move fetchCache import to test only path (fixes #928)
Although `fetchCache` is a conditional asynchronous import that's
never called outside of testing, some bundlers (like NextJS) will
try bundle it anyway, leading to build / compilation errors.

That's fair enough and to be safe we really should only import
this file from test code.  The new v3 architecture was designed
for this kind of setup, so, the quick fix below was quite easy
and we can still clean up the rest further at a later date.